% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brpl.R
\name{brpl}
\alias{brpl}
\title{Calculate Bivariate Quantiles}
\usage{
brpl(data, var1, var2, tau = 0.5, nalpha = 100)
}
\arguments{
\item{data}{Input data frame (tibbles and other data.frame variants are automatically converted to data.frame)}

\item{var1}{Name of first variable}

\item{var2}{Name of second variable}

\item{tau}{Quantile level (default: 0.5)}

\item{nalpha}{Number of alpha values (default: 100)}
}
\value{
An object of class brplPlot
}
\description{
Calculate Bivariate Quantiles
}
