% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_robject.R
\name{create_robject}
\alias{create_robject}
\title{Create a Custom Object of Class 'robject'}
\usage{
create_robject(variables)
}
\arguments{
\item{variables}{A numeric vector containing the variables to be stored in the object.}
}
\value{
An object of class \code{"robject"}.
}
\description{
This function creates a custom object of class \code{"robject"}, which stores a numeric vector of variables.
}
\details{
The \code{"robject"} class is a simple container for numeric data. It includes methods for printing
and plotting the stored data as a histogram.
}
