% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf_utils.R
\name{myecdf}
\alias{myecdf}
\title{Calculate Empirical Cumulative Distribution Function}
\usage{
myecdf(data, var, min.var = 0)
}
\arguments{
\item{data}{A data frame containing the variable}

\item{var}{Character string specifying the variable name}

\item{min.var}{Minimum value for the variable}
}
\value{
List containing ecdf and quantile functions
}
\description{
Calculate Empirical Cumulative Distribution Function
}
