% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutritionpoverty}
\alias{nutritionpoverty}
\title{Example dataset nutritionpoverty}
\format{
A data frame with 2,434 rows and 2 columns:
\describe{
\item{INval}{Income poverty value}
\item{NUval}{Nutrional value}
}
}
\usage{
nutritionpoverty
}
\description{
A subset of data from Franzi.
#TODO: Documentation
}
\keyword{datasets}
