% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pov_line_data_wt}
\alias{pov_line_data_wt}
\title{Poverty Line Data}
\format{
\subsection{\code{pov_line_data_wt}}{

A data frame with 30,221,284 rows and 4 columns:
\describe{
\item{leisure}{Weekly minutes per capita spent on leisure time. Defined as in Dorn et al. 2023.}
\item{inc_expenses}{Monthly income per capita. Defined as in Dorn et al. 2023.}
}
}
}
\source{
\url{https://www.inegi.org.mx/programas/enigh/nc/2018/}
}
\usage{
pov_line_data_wt
}
\description{
Poverty Line Data
}
\keyword{datasets}
