% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pov_line_example}
\alias{pov_line_example}
\title{Example dataset}
\format{
a \code{data.frame} with 67335 rows and 3 columns:
\describe{
\item{leisure}{Leisure time in minutes}
\item{inc_expenses}{Income based on expenditure share of households}
\item{weight}{Assumed population sampling weights}
}
}
\source{
Dorn, Franziska, et al. "A bivariate relative poverty line for leisure time and income poverty: Detecting intersectional differences using distributional copulas." Review of Income and Wealth 70.2 (2024): 395-419.
}
\usage{
pov_line_example
}
\description{
Dataset containing measurements for leisure time (in minutes) and income based on expenditure share
}
\keyword{datasets}
