% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prepquant}
\alias{prepquant}
\title{Main File for the wrapper}
\usage{
prepquant(data, tau, ecdfvar1, ecdfvar2, nalpha)
}
\arguments{
\item{data}{data source of interest. Has to be a data.frame.}

\item{tau}{Minimal possible value for given variable.}

\item{ecdfvar1}{ecdf object of variable 1 as constructed by myecdf function.}

\item{ecdfvar2}{ecdf object of variable 2 as constructed by myecdf function.}

\item{nalpha}{Minimal possible value for given variable.}
}
\description{
Main File for the wrapper
}
