% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{myecdf}
\alias{myecdf}
\title{Calculate the assumed CDF for the isoquant of the quantile boxes.}
\usage{
myecdf(data, var, min.var = 0)
}
\arguments{
\item{data}{data source of interest. Has to be a data.frame.}

\item{var}{Variable of interest. Has to be a character.}

\item{min.var}{Minimal possible value for given variable.
Defaults to 0 if not otherwise stated.}
}
\description{
Calculate the assumed CDF for the isoquant of the quantile boxes.
}
