% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_calculations.R
\name{prepquant}
\alias{prepquant}
\title{Prepare Quantile Calculations}
\usage{
prepquant(data, tau, ecdfvar1, ecdfvar2, nalpha)
}
\arguments{
\item{data}{Data frame with standardized variables}

\item{tau}{Quantile level}

\item{ecdfvar1}{ECDF for first variable}

\item{ecdfvar2}{ECDF for second variable}

\item{nalpha}{Number of alpha values}
}
\value{
Data frame with quantile calculations
}
\description{
Prepare Quantile Calculations
}
\keyword{internal}
